refactor(service patches): Automate service patches elsewhere
