+++ lib/ibm_watson/discovery_v1.rb
@@ -32,10 +32,6 @@ module IBMWatson
   ##
   # The Discovery V1 service.
   class DiscoveryV1
-    if File.exist?(Dir.getwd + "/lib/ibm_watson/service_extensions/patch_discovery_v1.rb")
-      require_relative("./service_extensions/patch_discovery_v1.rb")
-      DiscoveryV1.prepend DiscoveryV1Patch
-    end
     include Concurrent::Async
     ##
     # @!method initialize(args)
