+++ lib/ibm_watson/natural_language_understanding_v1.rb
@@ -36,10 +36,6 @@ module IBMWatson
   ##
   # The Natural Language Understanding V1 service.
   class NaturalLanguageUnderstandingV1
-    if File.exist?(Dir.getwd + "/lib/ibm_watson/service_extensions/patch_natural_language_understanding_v1.rb")
-      require_relative("./service_extensions/patch_natural_language_understanding_v1.rb")
-      NaturalLanguageUnderstandingV1.prepend NaturalLanguageUnderstandingV1Patch
-    end
     include Concurrent::Async
     ##
     # @!method initialize(args)
