+++ lib/ibm_watson/personality_insights_v3.rb
@@ -47,10 +47,6 @@ module IBMWatson
   ##
   # The Personality Insights V3 service.
   class PersonalityInsightsV3
-    if File.exist?(Dir.getwd + "/lib/ibm_watson/service_extensions/patch_personality_insights_v3.rb")
-      require_relative("./service_extensions/patch_personality_insights_v3.rb")
-      PersonalityInsightsV3.prepend PersonalityInsightsV3Patch
-    end
     include Concurrent::Async
     ##
     # @!method initialize(args)
