+++ lib/ibm_watson/tone_analyzer_v3.rb
@@ -38,10 +38,6 @@ module IBMWatson
   ##
   # The Tone Analyzer V3 service.
   class ToneAnalyzerV3
-    if File.exist?(Dir.getwd + "/lib/ibm_watson/service_extensions/patch_tone_analyzer_v3.rb")
-      require_relative("./service_extensions/patch_tone_analyzer_v3.rb")
-      ToneAnalyzerV3.prepend ToneAnalyzerV3Patch
-    end
     include Concurrent::Async
     ##
     # @!method initialize(args)
